Amended Derivative
The regular derivative is 53x faster. Now we append to this a boundary
term (in the non-hierarchical 1D position basis) c.f. Miller &
Schnetter. Next, we will transform to the hierarchical 1-D basis to get
a matrix form for this new derivative on hier space. This will
generalize readily to n-D via the tensor construction, and then we’ll
sparsify.